Update GraphQL Modules
