+++ packages/boost/package.json
@@ -27,7 +27,7 @@
     "@accounts/graphql-api": "^0.8.0",
     "@accounts/server": "^0.7.0",
     "@accounts/types": "^0.7.0",
-    "@graphql-modules/core": "^0.2.11",
+    "@graphql-modules/core": "^0.2.14",
     "apollo-server": "^2.2.3",
     "graphql": "^14.0.2",
     "graphql-tools": "^4.0.3",
