+++ packages/e2e/package.json
@@ -41,7 +41,7 @@
     "@accounts/rest-express": "^0.8.0",
     "@accounts/server": "^0.7.0",
     "@accounts/types": "^0.7.0",
-    "@graphql-modules/epoxy": "0.2.11",
+    "@graphql-modules/epoxy": "0.2.14",
     "@types/lodash": "4.14.118",
     "@types/mongoose": "5.3.1",
     "@types/node-fetch": "2.1.4",
