--- packages/graphql-api/src/utils/authenticated-directive.ts
@@ -1,14 +1,0 @@
-import { authenticated } from './authenticated-resolver';
-import { SchemaDirectiveVisitor } from 'graphql-tools';
-
-export class AuthenticatedDirective extends SchemaDirectiveVisitor {
-  public visitFieldDefinition(field: any) {
-    field.resolve = authenticated(field.resolve);
-  }
-  public visitObject(object: any) {
-    const fields = object.getFields();
-    Object.keys(fields).forEach((key: any) => {
-      fields[key].resolve = authenticated(fields[key].resolve);
-    });
-  }
-}
