--- packages/graphql-api/src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './authenticated-directive';
-export * from './authenticated-resolver';
